+++ CHANGELOG.md
@@ -1,5 +1,40 @@
 # Changelog
 
+## [5.6.0](https://github.com/npalm/terraform-aws-gitlab-runner/compare/5.5.0...5.6.0) (2022-12-11)
+
+
+### Features
+
+* allow setting runners.docker.services ([#491](https://github.com/npalm/terraform-aws-gitlab-runner/issues/491)) ([6d73e99](https://github.com/npalm/terraform-aws-gitlab-runner/commit/6d73e994c41bf3f6b53db4c940a7a013e534a685))
+* **asg:** Add fine-grained options for schedule_config scale_in and scale_out ([#586](https://github.com/npalm/terraform-aws-gitlab-runner/issues/586)) ([f72b8e3](https://github.com/npalm/terraform-aws-gitlab-runner/commit/f72b8e34c955c06b57ec1cd29714478d0568d71a))
+
+
+### Bug Fixes
+
+* Allow custom runner agent IAM role fixups ([#572](https://github.com/npalm/terraform-aws-gitlab-runner/issues/572)) ([#577](https://github.com/npalm/terraform-aws-gitlab-runner/issues/577)) ([bcb0c0e](https://github.com/npalm/terraform-aws-gitlab-runner/commit/bcb0c0eb9de74f0f67181cdeb8cb0fe91cb0e9b1))
+* runner_agent_role_arn ([#596](https://github.com/npalm/terraform-aws-gitlab-runner/issues/596)) ([b069b88](https://github.com/npalm/terraform-aws-gitlab-runner/commit/b069b88bf03557534ffa40bab6d21c085c09f021))
+
+## [5.5.0](https://github.com/npalm/terraform-aws-gitlab-runner/compare/5.4.1...5.5.0) (2022-11-27)
+
+
+### Features
+
+* Support volume type configuration ([#579](https://github.com/npalm/terraform-aws-gitlab-runner/issues/579)) ([b7dd834](https://github.com/npalm/terraform-aws-gitlab-runner/commit/b7dd8340a1c68221c77942fc6ac6a71c5eeb39fb))
+
+
+### Bug Fixes
+
+* Compress (zip) user_data to avoid max size ([#565](https://github.com/npalm/terraform-aws-gitlab-runner/issues/565)) ([64b8594](https://github.com/npalm/terraform-aws-gitlab-runner/commit/64b8594f3aa747a0cd23aafe9a37c051bbf51ceb))
+* ensure a complete `config.toml` before starting the GitLab Agent ([#574](https://github.com/npalm/terraform-aws-gitlab-runner/issues/574)) ([e32f3bc](https://github.com/npalm/terraform-aws-gitlab-runner/commit/e32f3bcec95df40fa655cab68e31da8a3ff8d090))
+* Use runners_pull_policies to set pull_policy instead of allowed_pull_policies ([#557](https://github.com/npalm/terraform-aws-gitlab-runner/issues/557)) ([a67b87b](https://github.com/npalm/terraform-aws-gitlab-runner/commit/a67b87b7e56eb76777133d011481586c2e0079a5))
+
+### [5.4.1](https://github.com/npalm/terraform-aws-gitlab-runner/compare/5.4.0...5.4.1) (2022-10-13)
+
+
+### Bug Fixes
+
+* Add var.environment to the name of the S3 cache bucket [#533](https://github.com/npalm/terraform-aws-gitlab-runner/issues/533) ([#555](https://github.com/npalm/terraform-aws-gitlab-runner/issues/555)) ([6e0cd97](https://github.com/npalm/terraform-aws-gitlab-runner/commit/6e0cd97e72e46bca8bb53493121f3eaf59526250))
+
 ## [5.4.0](https://github.com/npalm/terraform-aws-gitlab-runner/compare/5.3.0...5.4.0) (2022-10-10)